refactor(domain): 将比赛评分字段类型从Integer改为Long
- 修改MatchPlayerStats实体类中的rating字段类型为Long
- 更新MatchPlayerStatsDO数据对象中的rating字段类型为Long
- 更新MatchPlayerStatsDTO传输对象中的rating字段类型为Long
- 相应修改构造函数参数类型以匹配Long类型
- 保持所有相关getter和setter方法的一致性

+++ src/main/java/com/csxuhuan/gelatoni/application/dto/MatchPlayerStatsDTO.java
@@ -27,7 +27,7 @@ public class MatchPlayerStatsDTO {
     private String playerName;
 
     /** 比赛评分 */
-    private Integer rating;
+    private Long rating;
 
     /** 是否MVP：1=是 */
     private Boolean isMvp;
@@ -117,11 +117,11 @@ public class MatchPlayerStatsDTO {
         this.playerName = playerName;
     }
 
-    public Integer getRating() {
+    public Long getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(Long rating) {
         this.rating = rating;
     }
